# Conflicts:
#	core/src/main/java/com/orientechnologies/orient/core/db/document/ODatabaseDocumentTx.java
#	core/src/test/java/com/orientechnologies/orient/core/sql/OCommandExecutorSQLSelectTest.java

+++ core/src/main/java/com/orientechnologies/orient/core/metadata/sequence/OSequenceLibraryImpl.java
@@ -50,7 +50,7 @@ public class OSequenceLibraryImpl implements OSequenceLibrary {
 
   @Override
   public Set<String> getSequenceNames() {
-    return null;
+    return sequences.keySet();
   }
 
   @Override
@@ -85,8 +85,10 @@ public class OSequenceLibraryImpl implements OSequenceLibrary {
   public void dropSequence(String iName) {
     OSequence seq = getSequence(iName);
 
-    seq.getDocument().delete();
-    sequences.remove(iName);
+    if (seq != null) {
+      seq.getDocument().delete();
+      sequences.remove(iName);
+    }
   }
 
   @Override
